p_filter_query.pck
I saw something where a sqlserver stored procedure accepted xml as input and dynamically queried a given table based on this filter. This is a bit more involved with Oracle as a query is so strongly bound to the output column names and dbms_sql is required rather than execute immediate to emulate this. This will change in 12c the return result function https://oracle-base.com/articles/12c/implicit-statement-results-12cr1